HTTPS not HTTP and Key Info
Note that CMS operates entirely off of https and not http. Nginx will either throw an error or redirect you depending on the mood it's in.

Also, just wanted to say that the key is passed in an authorization header. I like how they are going to per project keys.